receive webapp event on session list
